+++ crates/core/tests/verbatim.rs
@@ -20,7 +20,7 @@ fn test_verbatim() {
     contract.derive_storage_pointers();
 
     // Get main bytecode with verbatim
-    match Codegen::generate_main_bytecode(&EVMVersion::default(), &contract, None) {
+    match Codegen::generate_main_bytecode(&EVMVersion::default(), &contract, None, false) {
         Ok(mb) => assert_eq!(mb, "1234567890abcdef".to_string()),
         Err(_) => panic!("moose"),
     }
@@ -42,5 +42,5 @@ fn test_verbatim_invalid_hex() {
     contract.derive_storage_pointers();
 
     // Expect failure to generate bytecode with verbatim
-    assert!(Codegen::generate_main_bytecode(&EVMVersion::default(), &contract, None).is_err());
+    assert!(Codegen::generate_main_bytecode(&EVMVersion::default(), &contract, None, false).is_err());
 }