+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 # Huff Neo Compiler changelog
 
 ## Unreleased
+- Introducing `--relax-jumps` CLI flag to optimize jump instructions from PUSH2 to PUSH1.
+  - Optimizes jump instructions from PUSH2 to PUSH1 when jump targets are 0-255.
+  - Reduces deployment gas costs, but will not affect runtime gas costs.
+- Fix parser bug where `for` loops and `if/else` statements after labels were rejected.
 
 ## [1.5.2] - 2025-11-05
 - Add compile-time if/else if/else statements.